requirements.txt
anyascii==0.3.1
APScheduler==3.6.3
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
bleach==3.3.1
cachetools==4.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.5
django-debug-toolbar==3.4.0
django-dotenv==1.4.2
django-extensions==3.1.5
django-filter==21.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-recaptcha==3.0.0
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==3.3
importlib-metadata==4.11.4
l18n==2021.3
Markdown==3.3.7
oauthlib==3.2.0
openpyxl==3.0.10
packaging==21.3
Pillow==9.1.1
polib==1.1.1
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-telegram-bot==13.12
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
telepath==0.2
tornado==6.1
typing-extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
wagtail==3.0
wagtail-django-recaptcha==1.0
wagtail-localize==1.2
wagtail-markdown==0.8.0
wagtail-metadata==4.0.0
wagtailcodeblock==1.28.0.0
webencodings==0.5.1
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
zipp==3.8.0
